refactor: split e2e_tests.py into per-area modules under automation/e2e_tests/
Replaces the single automation/e2e_tests.py with a package where each
UI area has its own module (builder, scrollbar, mana, buttons, widgets,
images, golden, launch). Shared infrastructure lives in common.py.
Entry point and python -m automation.e2e_tests invocation unchanged.

Closes #235
