+++ automation/e2e_tests/__main__.py
@@ -1,0 +1,5 @@
+import sys
+
+from automation.e2e_tests import main
+
+sys.exit(main())
